attempt to shirnk diff
