+++ crates/agentgateway/src/mcp/handler.rs
@@ -1,5 +1,5 @@
 use std::borrow::Cow;
-use std::collections::HashMap;
+use std::collections::{HashMap, HashSet};
 use std::sync::Arc;
 use std::sync::atomic::{AtomicU64, Ordering};
 
@@ -25,13 +25,13 @@ use crate::http::sessionpersistence::MCPSession;
 use crate::mcp;
 use crate::mcp::mergestream::{MergeFn, Messages};
 use crate::mcp::rbac::{CelExecWrapper, McpAuthorizationSet};
-use crate::mcp::router::McpBackendGroup;
-use crate::mcp::routing::{NameDecision, NameRouting};
+use crate::mcp::router::{McpBackendGroup, McpTarget};
 use crate::mcp::streamablehttp::{RequestProtocol, ServerSseMessage};
 use crate::mcp::upstream::{IncomingRequestContext, UpstreamError};
 use crate::mcp::{ClientError, FailureMode, MCPInfo, apps, mergestream, rbac, upstream};
 use crate::proxy::httpproxy::PolicyClient;
 use crate::telemetry::log::{AsyncLog, SpanWriteOnDrop, SpanWriter};
+use crate::types::agent::McpPrefixMode;
 
 fn rewrite_resource_messages(
 	routing: &NameRouting,
@@ -73,6 +73,130 @@ fn set_subscription_ack_id(
 		ack.params.meta = Some(meta);
 	}
 	message
+}
+
+const DELIMITER: &str = "_";
+
+/// How downstream-visible tool/prompt names map to upstream targets.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub enum NameRouting {
+	/// Single target: names pass through untouched and everything routes to it.
+	Single(String),
+	/// Names are exposed as `{target}_{name}`; the prefix is parsed off to route.
+	Prefix,
+	/// Names pass through untouched; calls are routed by looking up which
+	/// target serves the name. Requires names to be unique across targets.
+	Resolve,
+}
+
+impl NameRouting {
+	pub fn new(prefix_mode: McpPrefixMode, targets: &[Arc<McpTarget>]) -> Self {
+		if targets.len() != 1 {
+			match prefix_mode {
+				McpPrefixMode::Never => NameRouting::Resolve,
+				_ => NameRouting::Prefix,
+			}
+		} else if prefix_mode == McpPrefixMode::Always {
+			NameRouting::Prefix
+		} else {
+			NameRouting::Single(targets[0].name.to_string())
+		}
+	}
+
+	/// Whether the owning target must be resolved against upstream listings.
+	pub fn needs_resolution(&self) -> bool {
+		matches!(self, NameRouting::Resolve)
+	}
+
+	/// Whether resource URIs carry the target name (`{target}+{scheme}://...`).
+	/// Unlike names, URIs stay encoded even in Resolve mode: clients only ever
+	/// see URIs we produced, so the encoding is transparent to them.
+	pub fn encodes_uris(&self) -> bool {
+		!matches!(self, NameRouting::Single(_))
+	}
+
+	/// Downstream-visible form of an upstream tool/prompt name.
+	pub fn encode_name(&self, target: &str, name: &str) -> String {
+		match self {
+			NameRouting::Prefix => format!("{target}{DELIMITER}{name}"),
+			_ => name.to_string(),
+		}
+	}
+
+	/// Downstream-visible form of an upstream resource URI.
+	pub fn encode_uri(&self, target: &str, uri: &str) -> String {
+		if !self.encodes_uris() {
+			return uri.to_string();
+		}
+		// Apps UI resources must keep their ui:// scheme so hosts still
+		// recognize them; the target is carried in the authority instead.
+		if let Some(rewritten) = apps::encode_ui_uri(target, uri) {
+			return rewritten;
+		}
+		// Transform URI to service+scheme:// format for multiplexing
+		// e.g., "http://example.com" becomes "service+http://example.com"
+		if let Some(scheme_end) = uri.find("://") {
+			let (scheme, rest) = uri.split_at(scheme_end);
+			format!("{target}+{scheme}{rest}")
+		} else {
+			// URI must have a scheme - if not, return as-is and let validation handle it
+			uri.to_string()
+		}
+	}
+
+	/// Reverse of `encode_uri`: extract the (unvalidated) service name and the
+	/// original URI from a downstream URI. The caller is responsible for
+	/// validating the service name against the known targets.
+	pub fn decode_uri<'a, 'b: 'a>(
+		&'a self,
+		uri: &'b str,
+	) -> Result<(&'a str, String), UpstreamError> {
+		if let NameRouting::Single(default) = self {
+			return Ok((default.as_str(), uri.to_string()));
+		}
+		if apps::is_ui_uri(uri) {
+			return apps::decode_ui_uri(uri)
+				.ok_or_else(|| UpstreamError::InvalidRequest("invalid resource URI".to_string()));
+		}
+		// URI format: "service+scheme://rest"
+		let (service, original_uri) = uri
+			.split_once('+')
+			.ok_or_else(|| UpstreamError::InvalidRequest("invalid resource URI".to_string()))?;
+		// ui:// resources use the ui://service+rest namespace exclusively
+		if apps::is_ui_uri(original_uri) {
+			return Err(UpstreamError::InvalidRequest(
+				"invalid resource URI".to_string(),
+			));
+		}
+		Ok((service, original_uri.to_string()))
+	}
+
+	/// Names served by more than one target, which are unroutable in Resolve
+	/// mode and dropped from merged listings; empty for the other modes.
+	pub fn duplicate_names<'a>(
+		&self,
+		kind: &str,
+		names: impl Iterator<Item = (&'a Strng, &'a str)>,
+	) -> HashSet<String> {
+		if !self.needs_resolution() {
+			return HashSet::new();
+		}
+		let mut owners: HashMap<&str, Vec<&Strng>> = HashMap::new();
+		for (target, name) in names {
+			owners.entry(name).or_default().push(target);
+		}
+		owners
+			.into_iter()
+			.filter(|(_, targets)| targets.len() > 1)
+			.map(|(name, targets)| {
+				warn!(
+					"dropping {kind} '{name}': served by multiple targets ({})",
+					targets.iter().join(", "),
+				);
+				name.to_string()
+			})
+			.collect()
+	}
 }
 
 /// What kind of name is being resolved to a target (`prefixMode: never`).
@@ -200,26 +324,30 @@ impl Relay {
 		})
 	}
 
-	/// Statically decode a name's target, or `None` when the owning target can
-	/// only be found by resolving against upstream listings (`prefixMode: never`).
-	pub fn static_target<'a>(&'a self, name: &'a str) -> Result<Option<&'a str>, UpstreamError> {
-		match self.upstreams.name_routing.decode_name(name)? {
-			NameDecision::Target { service, .. } => Ok(Some(service)),
-			NameDecision::NeedsResolution => Ok(None),
-		}
+	/// Whether names carry no routing information (`prefixMode: never`), so the
+	/// owning target can only be found by listing upstreams.
+	pub fn needs_resolution(&self) -> bool {
+		self.upstreams.name_routing.needs_resolution()
 	}
 
-	/// Map a name to its target, resolving against upstream listings when the
-	/// name itself carries no routing information (`prefixMode: never`).
-	pub async fn resolve_resource_name<'a, 'b: 'a>(
+	/// Map a downstream name to its `(target, upstream_name)`. In Resolve mode
+	/// (`prefixMode: never`) the name carries no route, so the owner is found by
+	/// listing every target.
+	pub async fn parse_resource_name<'a, 'b: 'a>(
 		&'a self,
 		kind: ResolveKind,
 		res: &'b str,
 		ctx: &IncomingRequestContext,
 	) -> Result<(Cow<'a, str>, &'b str), UpstreamError> {
-		match self.upstreams.name_routing.decode_name(res)? {
-			NameDecision::Target { service, name } => Ok((Cow::Borrowed(service), name)),
-			NameDecision::NeedsResolution => {
+		match &self.upstreams.name_routing {
+			NameRouting::Single(default) => Ok((Cow::Borrowed(default.as_str()), res)),
+			NameRouting::Prefix => res
+				.split_once(DELIMITER)
+				.map(|(service, name)| (Cow::Borrowed(service), name))
+				.ok_or(UpstreamError::InvalidRequest(
+					"invalid resource name".to_string(),
+				)),
+			NameRouting::Resolve => {
 				let target = self.resolve_unprefixed(kind, res, ctx).await?;
 				Ok((Cow::Owned(target.to_string()), res))
 			},
@@ -278,18 +406,13 @@ impl Relay {
 			}
 		}
 
-		// Match the authorization-denied message so callers cannot probe which
-		// names exist. Collisions get the same treatment: naming the targets
-		// would leak topology, so they are only logged.
-		let denied = || UpstreamError::Authorization {
-			resource_type: kind.as_str().to_string(),
-			resource_name: name.to_string(),
-		};
 		match owners.as_slice() {
 			[one] => Ok(one.clone()),
 			// A skipped target might have served the name; surface the outage
 			// rather than reporting a possibly-existing name as unknown.
-			[] => Err(skipped.unwrap_or_else(denied)),
+			[] => Err(skipped.unwrap_or_else(|| {
+				UpstreamError::InvalidRequest(format!("unknown {} {}", kind.as_str(), name))
+			})),
 			many => {
 				warn!(
 					"{} '{}' is served by multiple targets ({}); names must be unique across targets when prefixMode is 'never'",
@@ -297,7 +420,11 @@ impl Relay {
 					name,
 					many.iter().join(", "),
 				);
-				Err(denied())
+				Err(UpstreamError::InvalidRequest(format!(
+					"{} {} is served by multiple targets",
+					kind.as_str(),
+					name
+				)))
 			},
 		}
 	}
